chore(suite): move SuiteLayout hooks into the component, remove refs from dependency arrays
